refactor: allow command logic to run in suspended function
`CliktCommand#run` is used by `CliktCommand#parse` to parse messages into respective args. The commands whose `#run` is invoked is saved in `Context#obj`. After the parsing, the list is retrieved and iterated on, calling `AbstractCommand#execute` on each.

This is done to work around the fact that `CliktCommand#run` does not have the `suspend` modifier.