feat: implement CLI-styled commands (#57)
* feat: implement a proof-of-concept CLI to improve UX

Proof-of-concept was done on health check commands first due to their small scope. The old implementation is replaced with their CLI counterparts.

* improvement: convert miscellaneous commands to CLI

* improvement: convert music commands to CLI

* improvement: improve log formatting and reduce redundant logs

* refactor: allow command logic to run in suspended function

`CliktCommand#run` is used by `CliktCommand#parse` to parse messages into respective args. The commands whose `#run` is invoked is saved in `Context#obj`. After the parsing, the list is retrieved and iterated on, calling `AbstractCommand#execute` on each.

This is done to work around the fact that `CliktCommand#run` does not have the `suspend` modifier.

* improvement: convert Shiritori commands to CLI

Music commands were renamed to improve clarity and avoid name conflicts.

* refactor: remove unused classes and reorganize package structure

* refactor: move game over listener logic to its own class

* bug: prevent a game from ending if a player times out in a game with more than 2 players